+++ tests/testthat.R
@@ -1,0 +1,3 @@
+library(testthat)
+library(Rttf2pt1)
+test_check("Rttf2pt1")